Flatten src/semantic/assembly/ into a single file
Consolidate `src/semantic/assembly/mod.rs` and `src/semantic/assembly/model.rs` into `src/semantic/assembly.rs` to reduce module depth and simplify the project structure, adhering to the "Razor" directives.

* Move contents of `model.rs` (structs and constants) to `assembly.rs`.
* Move contents of `mod.rs` (Assembler logic) to `assembly.rs`.
* Ensure `AssemblyError` is public for re-export.
* Delete `src/semantic/assembly/` directory.
